requirements.txt
httpx>=0.27
python-dotenv>=1.0
jupyter>=1.0
ipykernel>=6.0
pandas>=2.0
numpy>=1.26
tqdm>=4.66
matplotlib>=3.8
scikit-learn>=1.4
xgboost>=2.0
